gapic: refactor gengapic pkg
* gapic: refactor gengapic codebase
* bazel: run gazelle to update bazel files
