+++ .github/release-tool/BUILD.bazel
@@ -1,0 +1,1 @@
+# gazelle:ignore
