+++ cmd/protoc-gen-go_cli/BUILD.bazel
@@ -2,19 +2,19 @@ load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
 
 # gazelle:proto disable_global
 go_library(
-    name = "go_default_library",
+    name = "protoc-gen-go_cli_lib",
     srcs = ["main.go"],
     importpath = "github.com/googleapis/gapic-generator-go/cmd/protoc-gen-go_cli",
     visibility = ["//visibility:private"],
     deps = [
-        "//internal/gencli:go_default_library",
-        "@com_github_golang_protobuf//proto:go_default_library",
-        "@io_bazel_rules_go//proto/wkt:compiler_plugin_go_proto",
+        "//internal/gencli",
+        "@com_github_golang_protobuf//proto",
+        "@com_github_golang_protobuf//protoc-gen-go/plugin",
     ],
 )
 
 go_binary(
     name = "protoc-gen-go_cli",
-    embed = [":go_default_library"],
+    embed = [":protoc-gen-go_cli_lib"],
     visibility = ["//visibility:public"],
 )
