+++ internal/errors/BUILD.bazel
@@ -1,7 +1,7 @@
 load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
-    name = "go_default_library",
+    name = "errors",
     srcs = ["errors.go"],
     importpath = "github.com/googleapis/gapic-generator-go/internal/errors",
     visibility = ["//:__subpackages__"],
