+++ internal/gencli/BUILD.bazel
@@ -1,7 +1,7 @@
 load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
-    name = "go_default_library",
+    name = "gencli",
     srcs = [
         "cmd_file.go",
         "completion_file.go",
@@ -14,11 +14,11 @@ go_library(
     importpath = "github.com/googleapis/gapic-generator-go/internal/gencli",
     visibility = ["//:__subpackages__"],
     deps = [
-        "//internal/errors:go_default_library",
-        "//internal/pbinfo:go_default_library",
-        "//internal/printer:go_default_library",
+        "//internal/errors",
+        "//internal/pbinfo",
+        "//internal/printer",
         "@com_github_golang_protobuf//proto:go_default_library",
-        "@com_github_jhump_protoreflect//desc:go_default_library",
+        "@com_github_jhump_protoreflect//desc",
         "@go_googleapis//google/api:annotations_go_proto",
         "@io_bazel_rules_go//proto/wkt:compiler_plugin_go_proto",
         "@io_bazel_rules_go//proto/wkt:descriptor_go_proto",
@@ -26,7 +26,7 @@ go_library(
 )
 
 go_test(
-    name = "go_default_test",
+    name = "gencli_test",
     srcs = [
         "cmd_file_test.go",
         "completion_file_test.go",
@@ -35,10 +35,10 @@ go_test(
         "service_file_test.go",
     ],
     data = glob(["testdata/**"]),
-    embed = [":go_default_library"],
+    embed = [":gencli"],
     deps = [
-        "//internal/pbinfo:go_default_library",
-        "//internal/txtdiff:go_default_library",
+        "//internal/pbinfo",
+        "//internal/txtdiff",
         "@io_bazel_rules_go//proto/wkt:descriptor_go_proto",
     ],
 )
