+++ internal/gengapic/BUILD.bazel
@@ -1,56 +1,61 @@
 load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
-    name = "go_default_library",
+    name = "gengapic",
     srcs = [
         "client_init.go",
         "doc_file.go",
         "example.go",
+        "generator.go",
         "gengapic.go",
+        "helpers.go",
         "imports.go",
         "lro.go",
         "markdown.go",
+        "options.go",
         "paging.go",
         "stream.go",
     ],
     importpath = "github.com/googleapis/gapic-generator-go/internal/gengapic",
     visibility = ["//:__subpackages__"],
     deps = [
-        "//internal/errors:go_default_library",
-        "//internal/grpc_service_config:go_default_library",
-        "//internal/license:go_default_library",
-        "//internal/pbinfo:go_default_library",
-        "//internal/printer:go_default_library",
+        "//internal/errors",
+        "//internal/grpc_service_config",
+        "//internal/license",
+        "//internal/pbinfo",
+        "//internal/printer",
         "@com_github_golang_protobuf//proto:go_default_library",
-        "@com_gitlab_golang_commonmark_markdown//:go_default_library",
+        "@com_gitlab_golang_commonmark_markdown//:markdown",
         "@go_googleapis//google/api:annotations_go_proto",
         "@go_googleapis//google/api:serviceconfig_go_proto",
         "@go_googleapis//google/longrunning:longrunning_go_proto",
         "@go_googleapis//google/rpc:code_go_proto",
-        "@in_gopkg_yaml_v2//:go_default_library",
+        "@in_gopkg_yaml_v2//:yaml_v2",
         "@io_bazel_rules_go//proto/wkt:compiler_plugin_go_proto",
         "@io_bazel_rules_go//proto/wkt:descriptor_go_proto",
     ],
 )
 
 go_test(
-    name = "go_default_test",
+    name = "gengapic_test",
     srcs = [
         "client_init_test.go",
         "doc_file_test.go",
         "example_test.go",
         "gengapic_test.go",
+        "helpers_test.go",
         "markdown_test.go",
+        "options_test.go",
         "paging_test.go",
     ],
     data = glob(["testdata/**"]),
-    embed = [":go_default_library"],
+    embed = [":gengapic"],
     deps = [
-        "//internal/grpc_service_config:go_default_library",
-        "//internal/pbinfo:go_default_library",
-        "//internal/txtdiff:go_default_library",
+        "//internal/grpc_service_config",
+        "//internal/pbinfo",
+        "//internal/txtdiff",
         "@com_github_golang_protobuf//proto:go_default_library",
-        "@com_github_google_go_cmp//cmp:go_default_library",
+        "@com_github_google_go_cmp//cmp",
         "@go_googleapis//google/api:annotations_go_proto",
         "@go_googleapis//google/api:serviceconfig_go_proto",
         "@go_googleapis//google/longrunning:longrunning_go_proto",
@@ -58,6 +63,6 @@ go_test(
         "@io_bazel_rules_go//proto/wkt:descriptor_go_proto",
         "@io_bazel_rules_go//proto/wkt:duration_go_proto",
         "@io_bazel_rules_go//proto/wkt:wrappers_go_proto",
-        "@org_golang_google_protobuf//encoding/protojson:go_default_library",
+        "@org_golang_google_protobuf//encoding/protojson",
     ],
 )
