+++ internal/gensample/BUILD.bazel
@@ -1,7 +1,7 @@
 load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
-    name = "go_default_library",
+    name = "gensample",
     srcs = [
         "bytes.go",
         "enum.go",
@@ -20,19 +20,19 @@ go_library(
     importpath = "github.com/googleapis/gapic-generator-go/internal/gensample",
     visibility = ["//:__subpackages__"],
     deps = [
-        "//internal/errors:go_default_library",
-        "//internal/gensample/schema_v1p2:go_default_library",
-        "//internal/license:go_default_library",
-        "//internal/pbinfo:go_default_library",
-        "//internal/printer:go_default_library",
-        "@in_gopkg_yaml_v2//:go_default_library",
+        "//internal/errors",
+        "//internal/gensample/schema_v1p2",
+        "//internal/license",
+        "//internal/pbinfo",
+        "//internal/printer",
+        "@in_gopkg_yaml_v2//:yaml_v2",
         "@io_bazel_rules_go//proto/wkt:compiler_plugin_go_proto",
         "@io_bazel_rules_go//proto/wkt:descriptor_go_proto",
     ],
 )
 
 go_test(
-    name = "go_default_test",
+    name = "gensample_test",
     srcs = [
         "enum_test.go",
         "genmanifest_test.go",
@@ -42,15 +42,15 @@ go_test(
         "resource_test.go",
     ],
     data = glob(["testdata/**"]),
-    embed = [":go_default_library"],
+    embed = [":gensample"],
     deps = [
-        "//internal/errors:go_default_library",
-        "//internal/gensample/schema_v1p2:go_default_library",
-        "//internal/pbinfo:go_default_library",
-        "//internal/txtdiff:go_default_library",
+        "//internal/errors",
+        "//internal/gensample/schema_v1p2",
+        "//internal/pbinfo",
+        "//internal/txtdiff",
         "@com_github_golang_protobuf//proto:go_default_library",
-        "@com_github_google_go_cmp//cmp:go_default_library",
-        "@in_gopkg_yaml_v2//:go_default_library",
+        "@com_github_google_go_cmp//cmp",
+        "@in_gopkg_yaml_v2//:yaml_v2",
         "@io_bazel_rules_go//proto/wkt:descriptor_go_proto",
     ],
 )
