+++ internal/grpc_service_config/BUILD.bazel
@@ -22,7 +22,7 @@ go_proto_library(
 )
 
 go_library(
-    name = "go_default_library",
+    name = "grpc_service_config",
     srcs = ["config.go"],
     embed = [":grpc_service_config_go_proto"],
     importpath = "github.com/googleapis/gapic-generator-go/internal/grpc_service_config",
@@ -30,20 +30,20 @@ go_library(
     deps = [
         "@com_github_golang_protobuf//ptypes:go_default_library_gen",
         "@io_bazel_rules_go//proto/wkt:duration_go_proto",
-        "@org_golang_google_protobuf//encoding/protojson:go_default_library",
+        "@org_golang_google_protobuf//encoding/protojson",
     ],
 )
 
 go_test(
-    name = "go_default_test",
+    name = "grpc_service_config_test",
     srcs = ["config_test.go"],
-    embed = [":go_default_library"],
+    embed = [":grpc_service_config"],
     deps = [
-        "@com_github_google_go_cmp//cmp:go_default_library",
+        "@com_github_google_go_cmp//cmp",
         "@go_googleapis//google/rpc:code_go_proto",
         "@io_bazel_rules_go//proto/wkt:duration_go_proto",
         "@io_bazel_rules_go//proto/wkt:wrappers_go_proto",
-        "@org_golang_google_protobuf//encoding/protojson:go_default_library",
-        "@org_golang_google_protobuf//proto:go_default_library",
+        "@org_golang_google_protobuf//encoding/protojson",
+        "@org_golang_google_protobuf//proto",
     ],
 )
