+++ internal/printer/BUILD.bazel
@@ -1,7 +1,7 @@
 load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
-    name = "go_default_library",
+    name = "printer",
     srcs = ["printer.go"],
     importpath = "github.com/googleapis/gapic-generator-go/internal/printer",
     visibility = ["//:__subpackages__"],
