+++ internal/txtdiff/BUILD.bazel
@@ -1,7 +1,7 @@
 load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
-    name = "go_default_library",
+    name = "txtdiff",
     srcs = ["diff.go"],
     importpath = "github.com/googleapis/gapic-generator-go/internal/txtdiff",
     visibility = ["//:__subpackages__"],
